requirements.txt
a2wsgi==1.10.10
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-asgi-connector==1.1.2
aiosignal==1.4.0
aiosqlite==0.22.1
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
bentoml==1.4.35
blinker==1.9.0
cachetools==6.2.6
cattrs==23.1.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-option-group==0.5.9
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.8.2
idna==3.11
imbalanced-learn==0.14.1
imblearn==0.0
importlib_metadata==8.7.1
ipykernel==7.1.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kantoku==0.18.3
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
namex==0.1.0
narwhals==2.16.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.590.48
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opentelemetry-api==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-aiohttp-client==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==2.3.3
parso==0.8.5
pathspec==1.0.4
pexpect==4.9.0
pillow==12.1.0
pip-requirements-parser==32.0.1
platformdirs==4.5.1
plotly==6.5.2
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rich==14.3.2
rich-toolkit==0.18.1
rpds-py==0.30.0
schema==0.7.8
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
setuptools==80.10.2
simple-di==0.1.5
six==1.17.0
sklearn-compat==0.1.5
smmap==5.0.2
stack-data==0.6.3
starlette==0.52.1
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
toml==0.10.2
tomli_w==1.2.0
torch==2.10.0
tornado==6.5.4
traitlets==5.14.3
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.5.3
Werkzeug==3.1.5
wheel==0.46.3
wrapt==1.17.3
wsproto==1.3.2
xgboost==3.1.3
yarl==1.22.0
zipp==3.23.0
